requirements.txt
PyYAML
argparse
tqdm